# docs/marshaller_swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Marshaller API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/marshal/{serviceId}/{characteristicId}": {
            "post": {
                "description": "like POST /marshal, but the service is read from the device-repository by the id in the path and the characteristic comes from the path as well",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "marshal"
                ],
                "summary": "marshal a value into a protocol message, service and characteristic from the path",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of the service the message is for",
                        "name": "serviceId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "id of the characteristic the data is given in",
                        "name": "characteristicId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "the data to marshal",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/messages.MarshallingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "the protocol segments of the message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "the body is not valid json, or the service does not reference the given protocol",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
        "/unmarshal": {
            "post": {
                "description": "reads the message of the given service and returns its value in the requested characteristic",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "unmarshal"
                ],
                "summary": "unmarshal a protocol message into a value",
                "parameters": [
                    {
                        "description": "service, protocol and the message to unmarshal",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/messages.UnmarshallingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "the value in the requested characteristic",
                        "schema": {}
                    },
                    "400": {
                        "description": "the body is not valid json, or the service does not reference the given protocol",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/unmarshal/{serviceId}/{characteristicId}": {
            "post": {
                "description": "like POST /unmarshal, but the service is read from the device-repository by the id in the path and the characteristic comes from the path as well",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "unmarshal"
                ],
                "summary": "unmarshal a protocol message, service and characteristic from the path",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of the service the message came from",
                        "name": "serviceId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "id of the characteristic the value is wanted in",
                        "name": "characteristicId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "the message to unmarshal",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/messages.UnmarshallingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "the value in the requested characteristic",
                        "schema": {}
                    },
                    "400": {
                        "description": "the body is not valid json, or the service does not reference the given protocol",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
        "/v2/marshal/{serviceId}": {
            "post": {
                "description": "like POST /v2/marshal, but the service is read from the device-repository by the id in the path",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "marshal",
                    "v2"
                ],
                "summary": "marshal values into a protocol message, service from the path",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of the service the message is for",
                        "name": "serviceId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "the values to marshal",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/messages.MarshallingV2Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "the protocol segments of the message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "the body is not valid json, or the service does not reference the given protocol",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/v2/unmarshal": {
            "post": {
                "description": "returns the value at the requested path; without a path it is determined from the function and the requested aspects, closest aspect first",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "unmarshal",
                    "v2"
                ],
                "summary": "unmarshal a protocol message into a value",
                "parameters": [
                    {
                        "description": "service, protocol and the message to unmarshal",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/messages.UnmarshallingV2Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "the value in the requested characteristic",
                        "schema": {}
                    },
                    "400": {
                        "description": "the body is not valid json, the service does not reference the given protocol, or no output path matches the criteria",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/v2/unmarshal/{serviceId}": {
            "post": {
                "description": "like POST /v2/unmarshal, but the service is read from the device-repository by the id in the path",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "unmarshal",
                    "v2"
                ],
                "summary": "unmarshal a protocol message into a value, service from the path",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of the service the message came from",
                        "name": "serviceId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "the message to unmarshal",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/messages.UnmarshallingV2Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "the value in the requested characteristic",
                        "schema": {}
                    },
                    "400": {
                        "description": "the body is not valid json, the service does not reference the given protocol, or no output path matches the criteria",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        }
    },
    "definitions": {
        "configurables.Configurable": {
            "type": "object",
            "properties": {
                "characteristic_id": {
                    "type": "string"
                },
                "values": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/configurables.ConfigurableCharacteristicValue"
                    }
                }
            }
        },
        "configurables.ConfigurableCharacteristicValue": {
            "type": "object",
            "properties": {
                "label": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "messages.MarshallingRequest": {
            "type": "object",
            "properties": {
                "characteristic_id": {
                    "description": "semi-optional, may be determined by request path",
                    "type": "string"
                },
                "configurables": {
                    "description": "optional, may be empty",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/configurables.Configurable"
                    }
                },
                "data": {},
                "path_allow_list": {
                    "description": "optional\n\t\tif len \u003e 0: apply data only on given ContentVariable paths\n\t\tuseful if 2 variables have the same characteristic assigned but are used with different functions or aspects",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "protocol": {
                    "description": "semi-optional, may be determined by request path",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Protocol"
                        }
                    ]
                },
                "service": {
                    "description": "semi-optional, may be determined by request path",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Service"
                        }
                    ]
                }
            }
        },
        "messages.MarshallingV2Request": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.MarshallingV2RequestData"
                    }
                },
                "protocol": {
                    "description": "semi-optional, may be determined by service",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Protocol"
                        }
                    ]
                },
                "service": {
                    "description": "semi-optional, may be determined by request path",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Service"
                        }
                    ]
                }
            }
        },
...
        "messages.UnmarshallingRequest": {
            "type": "object",
            "properties": {
                "characteristic_id": {
                    "description": "semi-optional, may be determined by request path",
                    "type": "string"
                },
                "content_variable_hints": {
                    "description": "optional",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "path_allow_list": {
                    "description": "optional\n\t\tif len \u003e 0: apply data only on given ContentVariable paths\n\t\tuseful if 2 variables have the same characteristic assigned but are used with different functions or aspects",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "protocol": {
                    "description": "semi-optional, may be determined by service",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Protocol"
                        }
                    ]
                },
                "service": {
                    "description": "semi-optional, may be determined by request path",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Service"
                        }
                    ]
                }
            }
        },
        "messages.UnmarshallingV2Request": {
            "type": "object",
            "properties": {
                "aspect_node": {
                    "description": "deprecated: please use AspectNodes",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.AspectNode"
                        }
                    ]
                },
                "aspect_node_id": {
                    "description": "deprecated: please use AspectNodeIds",
                    "type": "string"
                },
                "aspect_node_ids": {
                    "description": "semi-optional, to determine AspectNodes if not set",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "aspect_nodes": {
                    "description": "semi-optional, to determine Path if not set, may themselves be determent by AspectNodeIds",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.AspectNode"
                    }
                },
                "characteristic_id": {
                    "description": "semi-optional, may be determined by request path",
                    "type": "string"
                },
                "function_id": {
                    "description": "semi-optional, to determine Path if not set",
                    "type": "string"
                },
                "message": {
                    "description": "semi-optional; may be needed to create serialized_output",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "path": {
                    "description": "semi-optional, may be determent by FunctionId and AspectNodes",
                    "type": "string"
                },
                "protocol": {
                    "description": "semi-optional, may be determined by service",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Protocol"
                        }
                    ]
                },
                "serialized_output": {
                    "description": "semi-optional; may be created from message",
                    "type": "object",
                    "additionalProperties": true
                },
                "service": {
                    "description": "semi-optional, may be determined by request path",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Service"
                        }
                    ]
                }
            }
        },
        "model.AspectNode": {
            "type": "object",
            "properties": {
                "ancestor_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "aspect_class_id": {
                    "type": "string"
                },
                "child_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "descendent_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "string"
                },
                "root_id": {
                    "type": "string"
                }
            }
        },
        "model.MarshallingV2RequestData": {
            "type": "object",
            "properties": {
                "aspect_node": {
                    "description": "deprecated: please use AspectNodes",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.AspectNode"
                        }
                    ]
                },
                "aspect_nodes": {
                    "description": "optional, to determine Paths if they are not set, may be empty if only FunctionId should be searched",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.AspectNode"
                    }
                },
                "characteristic_id": {
                    "type": "string"
                },
                "function_id": {
                    "description": "semi-optional, to determine Paths if they are not set",
                    "type": "string"
                },
                "paths": {
                    "description": "semi-optional, may be determent by FunctionId",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "value": {}
            }
        },
        "model.Protocol": {
            "type": "object",
            "properties": {
                "constraints": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "handler": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "protocol_segments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ProtocolSegment"
                    }
                }
            }
        },
        "model.Service": {
            "type": "object",
            "properties": {
                "attributes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Attribute"
                    }
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "inputs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Content"
                    }
                },
                "interaction": {
                    "$ref": "#/definitions/models.Interaction"
                },
                "local_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "outputs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Content"
                    }
                },
                "protocol_id": {
                    "type": "string"
                },
                "service_group_key": {
                    "type": "string"
                }
            }
        },
        "models.Attribute": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "origin": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "models.Content": {
            "type": "object",
            "properties": {
                "content_variable": {
                    "$ref": "#/definitions/models.ContentVariable"
                },
                "id": {
                    "type": "string"
                },
                "protocol_segment_id": {
                    "type": "string"
                },
                "serialization": {
                    "$ref": "#/definitions/models.Serialization"
                }
            }
        },
        "models.ContentVariable": {
            "type": "object",
            "properties": {
                "aspect_id": {
                    "description": "deprecated: please use AspectIds",
                    "type": "string"
                },
                "aspect_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "characteristic_id": {
                    "type": "string"
                },
                "function_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_void": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "omit_empty": {
                    "type": "boolean"
                },
                "serialization_options": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "sub_content_variables": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ContentVariable"
                    }
                },
                "type": {
                    "$ref": "#/definitions/models.Type"
                },
                "unit_reference": {
                    "type": "string"
                },
                "value": {}
            }
        },
...
        "models.Interaction": {
            "type": "string",
            "enum": [
                "event",
                "request",
                "event+request"
            ],
            "x-enum-varnames": [
                "EVENT",
                "REQUEST",
                "EVENT_AND_REQUEST"
            ]
        },
        "models.ProtocolSegment": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Serialization": {
            "type": "string",
            "enum": [
                "xml",
                "json",
                "plain-text"
            ],
            "x-enum-varnames": [
                "XML",
                "JSON",
                "PlainText"
            ]
        },
        "models.Type": {
            "type": "string",
            "enum": [
                "https://schema.org/Text",
                "https://schema.org/Integer",
                "https://schema.org/Float",
                "https://schema.org/Boolean",
                "https://schema.org/ItemList",
                "https://schema.org/StructuredValue"
            ],
            "x-enum-varnames": [
                "String",
                "Integer",
                "Float",
                "Boolean",
                "List",
                "Structure"
            ]
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}